backend/requirements.txt
fastapi
uvicorn
docker
networkx
langchain
langchain-openai
pydantic
python-multipart
requests
python-dotenv
PyJWT
python-jose[cryptography]
supabase
